+++ shared/src/main/java/com/pelmenstar/projktSens/shared/time/TimeConstants.java
@@ -30,7 +30,7 @@ public final class TimeConstants {
     public static final int SECONDS_IN_DAY = 24 * SECONDS_IN_HOUR;
 
     public static final int DAYS_PER_CYCLE = 146097;
-    public static final long DAYS_0000_TO_1970 = (DAYS_PER_CYCLE * 5) - (30 * 365 + 7);
+    public static final int DAYS_0000_TO_1970 = (DAYS_PER_CYCLE * 5) - (30 * 365 + 7);
 
     private TimeConstants() {}
 }
